% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanadvi.R
\name{stan_rdump_list}
\alias{stan_rdump_list}
\title{Dump a list of values to file}
\usage{
stan_rdump_list(l, file)
}
\arguments{
\item{l}{A named list containing data}

\item{file}{The output file}
}
\description{
This function dumps a named list of data to a file in the format
required by cmdStan, but accepting the data as a list (as in 
\code{stan}) rather than
variable names in the local environment (as in \code{stan_rdump}).
}

